(ios) Updating ElectrumConfigurationView with new warning about Tor
